+++ src/constants.py
@@ -1,6 +1,7 @@
 class Null:
 	__slots__ = ()
 
+
 TG_INPUT = "inp"  # tag input
 TG_OUTPUT = "out"  # tag output
 FLD_CONST = "const"  # const field